backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
requests==2.31.0
gunicorn==21.2.0